Refactor to use FileOperator to unzip file
