+++ app/src/main/java/com/appsubaruod/comicviewer/model/ComicModel.java
@@ -10,19 +10,11 @@ import com.appsubaruod.comicviewer.utils.messages.SetImageEvent;
 
 import org.greenrobot.eventbus.EventBus;
 
-import java.io.BufferedInputStream;
-import java.io.BufferedOutputStream;
 import java.io.File;
-import java.io.FileNotFoundException;
-import java.io.FileOutputStream;
-import java.io.IOException;
-import java.io.InputStream;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.concurrent.Executor;
 import java.util.concurrent.Executors;
-import java.util.zip.ZipEntry;
-import java.util.zip.ZipInputStream;
 
 /**
  * Created by s-yamada on 2017/03/28.
@@ -35,17 +27,35 @@ public class ComicModel {
     private Map<Integer, File> mFileMap = new HashMap<>();
 
     private final String LOG_TAG = "ComicModel";
-    private final Context mContext;
-    private final File mFileDir;
     private final Executor mWorkerThread = Executors.newSingleThreadExecutor();
 
-    private static final int BUFFER = 512;
-    private static final int TOOBIG = 0x6400000; // maximum file size : 100MB
-    private static final int TOOMANY = 10000;     // maximum file entries
+    private FileOperator mFileOperator;
+
+    private FileOperator.OnFileCopy mOnFileCopy = new FileOperator.OnFileCopy() {
+        @Override
+        public void onCopiedSingleFile(int fileCount, File outFile, int unpackedBytes) {
+            storeFileList(fileCount, outFile);
+            setMaxPageIndex(fileCount);
+
+            if (fileCount == 1) {
+                mPageIndex = 1;
+                // call postSticky, so as not to drop sending event during fragment transition
+                EventBus.getDefault().postSticky(new SetImageEvent(mPageIndex, obtainFile(mPageIndex)));
+                // notify book is opened
+                EventBus.getDefault().postSticky(new BookOpenedEvent());
+            }
+        }
+
+        @Override
+        public void onCopyCompleted(int maxFileCount) {
+            // Send notification including maxpage info
+            EventBus.getDefault().post(new LoadCompleteEvent(maxFileCount));
+        }
+    };
 
     private ComicModel(Context context) {
-        mContext = context;
-        mFileDir = mContext.getFilesDir();
+        mFileOperator = new FileOperator(context);
+        mFileOperator.registerCallback(mOnFileCopy);
     }
 
     public static ComicModel getInstance(Context mContext) {
@@ -68,10 +78,24 @@ public class ComicModel {
             @Override
             public void run() {
                 Log.d(LOG_TAG, uri.toString());
-                EventBus.getDefault().postSticky(new BookOpenedEvent());
-                unpackZip(uri);
+
+                copyToAppStorage(uri);
             }
         });
+    }
+
+    private void copyToAppStorage(Uri uri) {
+        initialize();
+        if (uri.getPath().contains(".zip")) {
+            // maybe zip file
+            mFileOperator.unpackZip(uri);
+        } else if (uri.getPath().endsWith(".png")
+                || uri.getPath().endsWith(".jpg")
+                || uri.getPath().endsWith(".jpeg")
+                || uri.getPath().endsWith(".gif")) {
+            // image file
+            mFileOperator.copyImageFiles(uri);
+        }
     }
 
     public void readNextPage() {
@@ -128,81 +152,9 @@ public class ComicModel {
         });
     }
 
-    private String validateFilename(String filename, String intendedDir) throws IOException {
-        File f = new File(filename);
-        String canonicalPath = f.getCanonicalPath();
 
-        File iD = new File(intendedDir);
-        String canonicalID = iD.getCanonicalPath();
-
-        if (canonicalPath.startsWith(canonicalID)) {
-            return canonicalPath;
-        } else {
-            throw new IllegalStateException("File is outside extraction target directory.");
-        }
-    }
 
     private void unpackZip(Uri zipUri) {
-
-        initialize();
-
-        InputStream is;
-        ZipInputStream zis = null;
-        try {
-            is = mContext.getContentResolver().openInputStream(zipUri);
-            zis = new ZipInputStream(new BufferedInputStream(is));
-            ZipEntry entry;
-            int entries = 0;
-            int total = 0;
-
-            while ((entry = zis.getNextEntry()) != null) {
-                System.out.println("Extracting: " + entry);
-                int count;
-                byte data[] = new byte[BUFFER];
-                // check if file name is valid and size is adequate
-                String name = validateFilename(entry.getName(), ".");
-                File outFile = new File(mFileDir + name);
-                // create parent dirs
-                outFile.getParentFile().mkdirs();
-                FileOutputStream fos = new FileOutputStream(outFile);
-                BufferedOutputStream dest = new BufferedOutputStream(fos, BUFFER);
-                while (total <= TOOBIG && (count = zis.read(data, 0, BUFFER)) != -1) {
-                    dest.write(data, 0, count);
-                    total += count;
-                }
-                dest.flush();
-                dest.close();
-                zis.closeEntry();
-
-                entries++;
-                storeFileList(entries, outFile);
-                setMaxPageIndex(entries);
-
-                if (entries == 1) {
-                    mPageIndex = 1;
-                    // call postSticky, so as not to drop sending event during fragment translation
-                    EventBus.getDefault().postSticky(new SetImageEvent(mPageIndex, obtainFile(mPageIndex)));
-                }
-                if (entries > TOOMANY) {
-                    throw new IllegalStateException("Too many files to unzip.");
-                }
-                if (total > TOOBIG) {
-                    throw new IllegalStateException("File being unzipped is too big.");
-                }
-            }
-        } catch (FileNotFoundException e) {
-            e.printStackTrace();
-        } catch (IOException e) {
-            e.printStackTrace();
-        } finally {
-            try {
-                zis.close();
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-            // Send notification including maxpage info
-            EventBus.getDefault().post(new LoadCompleteEvent(mMaxPageIndex));
-        }
     }
 
     private void initialize() {
